app/requirements.txt
numpy
paho-mqtt==1.5.1
pyinstaller==4.8